+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add metrics to calculate the number of blocks which fetched from the cache and fastnear
 * Add blocks chunks cache and optimize cache layer for read-rpc-server
 * Add authorization token supports
+* Migrate transaction details from GCS to ScyllaDB
+* Migrate `receipts_map` and `outcomes_map` from PostgreSQL to ScyllaDB
 
 ## [0.3.3](https://github.com/near/read-rpc/releases/tag/v0.3.3)